requirements.txt
albumentations>=1.1.0
comet-ml>=3.31.7
einops>=0.3.2
gradio>=2.9.4
imgaug==0.4.0
jsonargparse[signatures]>=4.3
kornia>=0.6.6
matplotlib>=3.4.3
omegaconf>=2.1.1
opencv-python
pandas
pytorch-lightning>=1.6.0,<1.7.0
timm==0.5.4
torchmetrics>=0.10.3
torchvision>=0.9.1,<=0.13.0
torchtext>=0.9.1,<=0.13.0
wandb==0.12.17
